part

+++ part3/coherenceRead.c
@@ -57,7 +57,7 @@ uint8_t* cacheSystemRead(cacheSystem_t* cacheSystem, uint32_t address, uint8_t I
 		if (shared == 0){
     
         }else{
-            uint8_t newID = returnFirstCacheID(cacheSystem->snooper, oldAddress, cacheSystem->blockDataSize);
+            int newID = returnFirstCacheID(cacheSystem->snooper, oldAddress, cacheSystem->blockDataSize);
             if (newID+1 != 0){
                 otherCache = getCacheFromID(cacheSystem,newID);
                 updateState(otherCache,oldAddress,INVALID);
